[atlas-gradle-plugin] upgrade to 2.2.3.rc4

+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/tools/diff/DiffResExtractor.java
@@ -359,7 +359,9 @@ public class DiffResExtractor {
 
             //copy file
             File rawFile = new File(fullResDir, diffFile);
-            FileUtils.copyFile(rawFile, new File(destDir, diffFile));
+            if (rawFile.exists()) {
+                FileUtils.copyFile(rawFile, new File(destDir, diffFile));
+            }
 
         }
